[Core] Move URLOutputStream to formatters package
The URLOutputStream is only used by formatters. Moving it to the
formatters package allows the visibility of the stream to be reduced
tidying up the effectively public API just a bit more.
