+++ core/src/main/java/cucumber/runtime/formatter/TestNGFormatter.java
@@ -13,8 +13,6 @@ import cucumber.api.event.TestSourceRead;
 import cucumber.api.event.TestStepFinished;
 import cucumber.api.formatter.StrictAware;
 import cucumber.runtime.CucumberException;
-import cucumber.runtime.io.URLOutputStream;
-import cucumber.runtime.io.UTF8OutputStreamWriter;
 import org.w3c.dom.Document;
 import org.w3c.dom.Element;
 import org.w3c.dom.NamedNodeMap;
