+++ core/src/main/java/cucumber/runtime/formatter/UTF8OutputStreamWriter.java
@@ -1,0 +1,11 @@
+package cucumber.runtime.formatter;
+
+import java.io.OutputStream;
+import java.io.OutputStreamWriter;
+import java.nio.charset.Charset;
+
+class UTF8OutputStreamWriter extends OutputStreamWriter {
+    UTF8OutputStreamWriter(OutputStream out) {
+        super(out, Charset.forName("UTF-8"));
+    }
+}
